created useSamplePreload hook to handle sample preload logic outside of context
